Move bootlaoder_hal.h to the openlcb directory. Clean up conditional compilation.
